requirements.txt
Markdown==3.4.1
Django==4.1
django-extensions==3.2.0
django-taggit==3.0.0